+++ sciety_labs/app/app_providers_and_models.py
@@ -18,7 +18,7 @@ from sciety_labs.providers.article_recommendation import (
 from sciety_labs.providers.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
-from sciety_labs.providers.async_providers.crossref.async_crossref import (
+from sciety_labs.providers.async_providers.crossref.providers import (
     AsyncCrossrefMetaDataProvider
 )
 from sciety_labs.providers.async_opensearch_article_recommendation import (
