+++ tests/unit_tests/app/routers/conftest.py
@@ -5,7 +5,7 @@ import pytest
 from sciety_labs.providers.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
-from sciety_labs.providers.async_providers.crossref.async_crossref import (
+from sciety_labs.providers.async_providers.crossref.providers import (
     AsyncCrossrefMetaDataProvider
 )
